refactor: refactor db init data path
